// src/app/Models/animales.ts
export interface Animales {
  siniiga: number;
  especie: string;
  genero: string;
  peso: number;
  altura: number;
  color: string;
  fechaNacimiento: Date;
  idCorral: number;
}

export interface AnimalesID {
  id: string;
  siniiga: number;
  especie: string;
  genero: string;
  peso: number;
  altura: number;
  color: string;
  fechaNacimiento: Date;
  idCorral: number;
}